[OPT] Optimize date picker related code, make the result key a public variable, alter some comment
